Add the ability to play multiple sounds at the same time
Add HashMaps in MediaBrowserService for MediaPlayer, the sound ArrayLists and the current sound. Use custom actions in the mediaSession callback to make it possible to specify the sound of the action. The service now sets the PlaybackState and the Playing Sound updates the isPlaying LiveData to update the play/pause Button